python/whisper-service/requirements.txt
# Whisper Service Dependencies
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI Whisper
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0

# File handling
aiofiles==23.2.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
